+++ hrushchev.nikita/FA/arithmetic.hpp
@@ -3,11 +3,11 @@
 
 namespace hrushchev
 {
-  long double sum(long long a, long long b);
-  long double subtract(long long a, long long b);
-  long double multiply(long long a, long long b);
-  long double division(long long a, long long b);
-  long double remaind(long long a, long long b);
+  long double sum(long double a, long double b);
+  long double subtract(long double a, long double b);
+  long double multiply(long double a, long double b);
+  long double division(long double a, long double b);
+  long long remaind(long long a, long long b);
   long double sin(double a);
   long double cos(double a);
 }
